+++ app/src/main/java/com/example/risuto/domain/model/SearchAnime.kt
@@ -11,7 +11,7 @@ data class SearchAnime(
         val synopsis : String,
         val type : String,
         val episodes : Int,
-        val score : Float,
+        val score : Double,
         val start_date : String,
         val end_date : String,
         val members : Int,
