+++ app/src/main/java/com/example/risuto/domain/model/TopAnime.kt
@@ -11,5 +11,5 @@ data class TopAnime(
     val start_date: String,
     val end_date: String,
     val members: Int,
-    val score: Float
+    val score: Double
 )
