+++ app/src/main/java/com/example/risuto/presentation/model/AnimeListPresentation.kt
@@ -1,12 +1,12 @@
 package com.example.risuto.presentation.model
 
 data class AnimeListPresentation(
-    val mal_id : Int,
-    val image_url : String,
-    val title : String,
-    val synopsis : String,
-    val type : String,
-    val episodes : Int,
-    val score : Float,
-    val members : Int
+    val mal_id: Int,
+    val image_url: String,
+    val title: String,
+    val synopsis: String,
+    val type: String,
+    val episodes: Int,
+    val score: Double,
+    val members: Int
 )
